requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
langchain
langchain-openai
langgraph